LayoutTests/imported/w3c/web-platform-tests/svg-aam/name/comp_labelledby-expected.txt

SVG-AAM: Labelledby Tests

Tests SVG-specific aria-labelledby rules in SVG-AAM รยง8.1 Name and Description, but note the open issues in SVG-AAM #31. Huey Dewey Louie Scrooge

SVG a[aria-labelledby]



SVG a[aria-label][aria-labelledby]



SVG a[aria-labelledby="{[aria-label]} {[aria-label]} {[xlink:title]} {* > [title]}"]




PASS [aria-labelledby] > circle
PASS [aria-labelledby] > rect
PASS [aria-labelledby] > polygon
PASS [aria-labelledby] > g
PASS [aria-label][aria-labelledby] > circle
PASS [aria-label][aria-labelledby] > rect
PASS [aria-label][aria-labelledby] > polygon
PASS [aria-label][aria-labelledby] > g
FAIL [aria-labelledby="{[aria-label]} {[xlink:title]} {* > [title]}"] > g assert_equals: <a href="#" aria-labelledby="nolan gilliam kaufman villeneuve" data-expectedlabel="Nolan Gilliam Kaufman Villeneuve" xlink:title="group link label" data-testname="[aria-labelledby=&quot;{[aria-label]} {[xlink:title]} {* &gt; [title]}&quot;] &gt; g" class="ex">
  <g fill="white" stroke="green" stroke-width="5">
    <circle cx="40" cy="40" r="25"></circle>
    <circle cx="60" cy="60" r="25"></circle>
  </g>
  </a> expected "Nolan Gilliam Kaufman Villeneuve" but got "Nolan Gilliam"
